[fix] 앱 이용방법 다이얼로그에서 첫 Fragment의 text가 보이지 않는 버그 수정
