+++ app/src/main/res/layout/activity_how_to_use.xml
@@ -12,7 +12,8 @@
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/how_to_use"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="0dp"
+            app:layout_constraintBottom_toTopOf="@id/indicator"
             app:layout_constraintTop_toTopOf="parent" />
 
         <com.google.android.material.tabs.TabLayout
